Rename Daypicker folder
