+++ src/elements/index.js
@@ -8,7 +8,7 @@ export { default as Notification } from './Notification';
 export { default as Typography } from './Typography';
 export { default as DatePicker } from './DatePicker';
 export { default as Icon } from './Icon';
-export { default as SingleDayPicker } from './SingleDayPicker';
+export { default as DayPicker } from './DayPicker';
 export { default as Accordion } from './Accordion';
 export { default as Progress } from './Progress';
 export { default as Card } from './Card';
